projects/이미지 찾아서 다운받기/requirements.txt
reportlab>=4.4.10
python-docx>=1.1.2
